requirements.txt
openpyxl==3.1.5
psutil==6.0.0
python-dotenv==1.0.1
